refactor: mybatis 설정 수정 및 추가
