+++ src/main/java/com/airjnc/common/config/MybatisConfig.java
@@ -12,7 +12,7 @@ import org.springframework.context.annotation.Configuration;
 
 @Configuration
 @RequiredArgsConstructor
-@MapperScan(basePackages = "com.airjnc.*.dao.mapper") // @Mppaer 스캔 경로
+@MapperScan(basePackages = "com.airjnc.*.dao") // @Mppaer 스캔 경로
 public class MybatisConfig {
 
   private final ApplicationContext context;
@@ -30,7 +30,11 @@ public class MybatisConfig {
     factoryBean.setDataSource(dataSource);
     factoryBean.setTypeAliasesPackage("com.airjnc.*.domain"); // 매핑되는 자바 객체 경로 설정
     factoryBean.setMapperLocations(
-        context.getResources("classpath:mybatis/mapper/*.xml")); // mapper 파일 경로 설정
+        context.getResources("classpath:mybatis/mapper/*.xml")
+    ); // mapper 파일 경로 설정
+    org.apache.ibatis.session.Configuration configuration = new org.apache.ibatis.session.Configuration();
+    configuration.setMapUnderscoreToCamelCase(true);
+    factoryBean.setConfiguration(configuration);
     return factoryBean.getObject();
   }
 }
