+++ src/main/resources/mybatis/mapper/user.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.airjnc.user.dao.mapper.UserMapper">
+<mapper namespace="com.airjnc.user.dao.UserMapper">
   <delete id="remove">
     UPDATE user
     set deleted_at = now()
