+++ src/test/java/com/airjnc/user/dao/impl/MybatisUserRepositoryTest.java
@@ -6,7 +6,7 @@ import static org.mockito.BDDMockito.then;
 import static org.mockito.BDDMockito.times;
 import com.airjnc.common.service.CommonCheckService;
 import com.airjnc.user.dao.UserRepository;
-import com.airjnc.user.dao.mapper.UserMapper;
+import com.airjnc.user.dao.UserMapper;
 import com.airjnc.user.domain.Gender;
 import com.airjnc.user.domain.UserEntity;
 import com.airjnc.user.dto.request.CreateDTO;
